fix: refactor diagnosis_report
