+++ test/unit/model/test_diagnosis_report.py
@@ -1,6 +1,6 @@
 import unittest
 
-from MIABIS_on_FHIR.MoF_diagnosis_report import DiagnosisReport
+from MIABIS_on_FHIR.diagnosis_report import DiagnosisReport
 
 
 class TestDiagnosisReport(unittest.TestCase):
